Move Color to its own file and StyleBuilder to its own DslMarker
